Common Assets/Template Project
Added a common folder to the root of the repository, it has a template project with the most updated ML package, MLTools package, Sdk, done in Unity Editor 2019.2.19f1.
The project also has a general asset folder structure, and a template scene with a general folder organization as well.
All of this will help provide consistency across all of our projects