Json testing
I made a test scene with two test scripts. One is a serializable script and the seccond assigns things and prints it to the console.

It's not doing exactly as expected but that might be because I wasn't running it in the simulator. It still kinda did as expected, but only for one pass.