Common Resources Update
Added SDK 0.24.1 and a new template project to the common resources folder. we know have the packages and template projects to work with either SDK 0.23 or 0.24, with the last unity versions that work with each. Also I added the latest version of MagicLeapTools to the newest project, though I do not believe there was any significant changes.